Most of these updates were done to fix compatability on the swith to REAL(10).
Most of the changes in $cmfdist/plane fix a variable name (LOG back to dLOG) that was
   inavertantly changed with the change of the inrtinsic function DLOG.

A few routines were still not updated from REAL*8.

There was bug fix to new_main/subs/var_x_opa_eta_v4.f (bad parallelization).

+++ new_main/subs/non_therm/get_lelec.f90
@@ -6,13 +6,13 @@
 	REAL(10) XKT(NKT)
 	REAL(10) N_ELEC
 !
-	real*8, parameter :: ELECTRON_VOLT=1.60217733D-12  ! erg
-	real*8, parameter :: PLANCKS_CONSTANT=6.626075D-27 ! erg sec
-	real*8, parameter :: SPEED_OF_LIGHT=2.99792458D+10 ! cm / sec
-	real*8, parameter :: PI = 3.141592653589793238462643D0
-	real*8, parameter :: ELECTRON_MASS=9.109389D-28    !gm
-	real*8, parameter :: ELECTRON_CHARGE = 4.803206814D-10 ! esu
-	real*8, parameter :: a0 = 0.529189379D-8    ! Bohr radius in cm
+	real(10), parameter :: ELECTRON_VOLT=1.60217733D-12  ! erg
+	real(10), parameter :: PLANCKS_CONSTANT=6.626075D-27 ! erg sec
+	real(10), parameter :: SPEED_OF_LIGHT=2.99792458D+10 ! cm / sec
+	real(10), parameter :: PI = 3.141592653589793238462643D0
+	real(10), parameter :: ELECTRON_MASS=9.109389D-28    !gm
+	real(10), parameter :: ELECTRON_CHARGE = 4.803206814D-10 ! esu
+	real(10), parameter :: a0 = 0.529189379D-8    ! Bohr radius in cm
 !
 	INTEGER I
 	INTEGER IKT
